FIA-138/FIA-152: Document deployment modes and centralize Docker fixtures
