+++ docs/testing.md
@@ -3,6 +3,10 @@
 TradeBot uses a test pyramid with explicit commands and explicit safety gates.
 Nox is the canonical interface for routine local validation and CI.
 
+For how these test layers map to local in-process, local Docker, automated
+Docker integration, and future Kubernetes deployment modes, see
+`docs/deployment.md`.
+
 Raw `python -m pytest` can still be useful as a tight local debugging escape
 hatch, but shared project workflows should use Nox so local and CI behavior
 stay aligned.
