+++ tests/integration/docker/conftest.py
@@ -1,0 +1,1 @@
+pytest_plugins = ("tests.fixtures.docker_service_stack",)
