+++ tests/integration/docker/test_quotes_service_stack.py
@@ -1,20 +1,8 @@
-import os
-
 import httpx
 import pytest
 
 
 pytestmark = [pytest.mark.service, pytest.mark.docker, pytest.mark.integration]
-
-QUOTES_BASE_URL_ENV_VAR = "TRADEBOT_TEST_QUOTES_BASE_URL"
-
-
-@pytest.fixture
-def quotes_base_url() -> str:
-    base_url = os.getenv(QUOTES_BASE_URL_ENV_VAR)
-    if not base_url:
-        pytest.skip(f"set {QUOTES_BASE_URL_ENV_VAR} to run quotes service stack tests")
-    return base_url.rstrip("/")
 
 
 def test_quotes_service_returns_simulator_backed_equity_quote(quotes_base_url: str):
